LPS-X Merge InitialUpgradeExtender into LiferayServiceExtender to avoid creating InitialUpgradeStep as DM component.
